setting the shaders subdir is data is avalable
This will allows for the shaders to be compiled/verified when the rest of the code is compiled. This pervents a lot problems since syntax errors in the shader files can be found during compile time and there is no longer a need to wait for the program to load them before exiting with the error message.