improve simulator loading
This commit improves the loading process of the simulaor. Now it is no longer possible to double click the start button and cause a crash. In addition to that a loading page is visible when while the simulator is loading its assets. This should make for a better experience using the simulator.